+++ src/components/admin/table-list/TableList.vue
@@ -15,7 +15,7 @@ import TableListNavSearch from './table-list-nav/TableListNavSearch.vue'
 import TableListNavPagination from './table-list-nav-pagination/TableListNavPagination.vue'
 import TableListNavFilter from './table-list-nav-filter/TableListNavFilter.vue'
 import SkeletonTable from '../../ui/skeleton-table/SkeletonTable.vue'
-import type { ITableListConfig } from './types/ITableListConfig'
+import type { ITableListConfig } from 'src/types'
 import type { TApiData } from 'src/types/IApiResource'
 import Card from '../../ui/card/Card.vue'
 import TableListNavSortable from './table-list-nav/TableListNavSortable.vue'
