+++ src/components/admin/titlebar/Titlebar.vue
@@ -3,12 +3,7 @@ import IconButton from '../../ui/icon-button/IconButton.vue'
 import Dropdown from '../../ui/dropdown/Dropdown.vue'
 import DropdownItemButton from '../../ui/dropdown/DropdownItemButton.vue'
 import Button from '../../ui/button/Button.vue'
-
-export interface IAction {
-	class?: string
-	label: string
-	onAction(): void
-}
+import type { IAction } from 'src/types/IAction'
 
 defineProps<{
 	title?: string | null
