+++ src/components/ui/aside/AsideSection.vue
@@ -1,11 +1,9 @@
 <script setup lang="ts">
+import type { IAction } from 'src/types/IAction'
 import Link from '../link/Link.vue'
 defineProps<{
 	title?: string
-	actions?: Array<{
-		label: string
-		onAction(): void
-	}>
+	actions?: IAction[]
 }>()
 </script>
