+++ src/types/IAction.ts
@@ -1,0 +1,11 @@
+import type { IVariant } from 'src/types/ITypes'
+
+export interface IAction {
+	leadingIcon?: string
+	label?: string
+	onAction?: any
+	form?: string
+	type?: string
+	variant?: IVariant
+	to?: object
+}
