+++ src/types/ITypes.ts
@@ -1,0 +1,1 @@
+export type IVariant = 'primary' | 'secondary' | 'success' | 'danger' | 'link' | 'dark' | 'plain'
